Moved new logs HTTP API under /manager
so http://.../api/logs is now http://.../api/manager/logs
