+++ rust/agama-server/src/web/service.rs
@@ -107,8 +107,7 @@ impl MainServiceBuilder {
                 state.clone(),
             ))
             .route("/ping", get(super::http::ping))
-            .route("/auth", post(login).get(session).delete(logout))
-            .nest("/logs", super::http::logs_router());
+            .route("/auth", post(login).get(session).delete(logout));
 
         tracing::info!("Serving static files from {}", self.public_dir.display());
         let serve = ServeDir::new(self.public_dir).precompressed_gzip();
